requirements.txt
groq
speech-recognition
AppOpener
edge-tts
pygame
python-dotenv
